# reduce garbage created while processing facets