@mschoch 2017-03-02
I forgot to include the perf numbers.  This comes from a benchmark on beer-search where we do a MatchAll query with facet on "abv" field.

```
$ benchcmp before.txt after.txt 
benchmark                 old ns/op     new ns/op     delta
BenchmarkTermFacets-4     237352244     163180908     -31.25%

benchmark                 old allocs     new allocs     delta
BenchmarkTermFacets-4     3726083        1881346        -49.51%

benchmark                 old bytes     new bytes     delta
BenchmarkTermFacets-4     152717395     79796837      -47.75%
```